remove tqdm add a progress bar for notebook move predict from core to api ignore inactive neurons
